SECOND

+++ src/main/java/com/usharesoft/basicmaven/Calculator.java
@@ -5,11 +5,11 @@ public class Calculator {
     public Calculator() {
     }
 
-    public int add(int numberOne, int numberTwo) {
+    public float add(int numberOne, int numberTwo) {
         return numberOne + numberTwo;
     }
 
-    public int subtract(int numberOne, int numberTwo) {
+    public float subtract(int numberOne, int numberTwo) {
         return numberOne - numberTwo;
     }
 
@@ -17,7 +17,7 @@ public class Calculator {
         return (float) numberOne / (float) numberTwo;
     }
     
-    public int multiply(int numberOne, int numberTwo) {
+    public float multiply(int numberOne, int numberTwo) {
         return numberOne * numberTwo;
     }
